requirements.txt
autopep8==1.7.0
certifi==2024.2.2
cfgv==3.3.1
charset-normalizer==2.1.1
decorator==4.4.2
distlib==0.3.6
ffmpeg==1.4
filelock==3.8.0
fitparse==1.2.0
identify==2.5.5
idna==3.4
imageio==2.31.0
imageio-ffmpeg==0.4.8
moviepy==1.0.3
nodeenv==1.7.0
numpy==1.24.2
pandas==1.5.0
Pillow==10.2.0
platformdirs==2.5.2
pre-commit==2.20.0
proglog==0.1.10
protobuf==4.21.6
pycodestyle==2.9.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.31.0
six==1.16.0
toml==0.10.2
tqdm==4.65.0
urllib3==2.2.1
virtualenv==20.16.5
opencv-python==4.9.0.80
garmin-fit-sdk==21.115.0

ed25519~=1.5
python-dotenv~=1.0.1
torch~=2.1.0.dev20230413
websocket~=0.2.1
gpxpy~=1.6.2
